Improve double click detection
Resolves issue #6.

Implemented using a hash map to store the last time a pose occurred. The
map uses the result of the pose's `toString()`. Using the pose object
itself as the key results in a compiler error. This could probably be
fixed by somehow telling the map to use toString() automatically when
hashing.

Note: this hasn't been tested yet as I'm on a plane and performing the
Myo's sync gesture would probably poke the person next to me.