Add tests that do not use MyoSimulator
MyoSimulator is great at testing the feature tree in the way most
similar to how it will be used (it will playback the events with the
appropriate delay between events). But in doing so it tests too much.
That is why tests that do not use MyoSimulator are added. They do not
test the entire feature tree, but instead test only the feature being
tested.